Реализовал функцию изменения рейтинга для объектов; Реализовал изменение рейтинга для записи; Удалил таблицу records_simple из БД

+++ SocialNetwork/Models/ratings_to_objects_with_rating.cs
@@ -17,7 +17,7 @@ namespace SocialNetwork.Models
         public int id { get; set; }
         public int user_id_from { get; set; }
         public int object_id { get; set; }
-        public byte value { get; set; }
+        public int value { get; set; }
     
         public virtual objects objects { get; set; }
         public virtual users users { get; set; }
